ONEM-30816: WPE 2.38 - User script is not visible in web inspector
