+++ LayoutTests/userscripts/window-onerror-for-isolated-world-1-expected.txt
@@ -2,13 +2,13 @@ Test that window.onerror and "error" event listeners from main world are invoked
 
 Main world window.onerror: Error: Error in main world inline script. at window-onerror-for-isolated-world-1.html:55:68 Error: Error in main world inline script.
 Main world error event listener: Error: Error in main world inline script. at window-onerror-for-isolated-world-1.html:55:68 Error: Error in main world inline script.
-Main world window.onerror: Error: Error in user script inline script. at undefined:12:20 null
-Main world error event listener: Error: Error in user script inline script. at undefined:12:20 null
+Main world window.onerror: Error: Error in user script inline script. at user-script:13:12:68 Error: Error in user script inline script.
+Main world error event listener: Error: Error in user script inline script. at user-script:13:12:68 Error: Error in user script inline script.
 Main world window.onerror: Error: Error in main world load handler. at window-onerror-for-isolated-world-1.html:51:72 Error: Error in main world load handler.
 Main world error event listener: Error: Error in main world load handler. at window-onerror-for-isolated-world-1.html:51:72 Error: Error in main world load handler.
-Main world window.onerror: Error: Error in user script load handler. at undefined:8:24 null
-Main world error event listener: Error: Error in user script load handler. at undefined:8:24 null
+Main world window.onerror: Error: Error in user script load handler. at user-script:13:8:72 Error: Error in user script load handler.
+Main world error event listener: Error: Error in user script load handler. at user-script:13:8:72 Error: Error in user script load handler.
 Main world window.onerror: Error: Error in main world setTimeout callback. at window-onerror-for-isolated-world-1.html:49:83 Error: Error in main world setTimeout callback.
 Main world error event listener: Error: Error in main world setTimeout callback. at window-onerror-for-isolated-world-1.html:49:83 Error: Error in main world setTimeout callback.
-Main world window.onerror: Error: Error in user script setTimeout callback. at undefined:6:28 null
-Main world error event listener: Error: Error in user script setTimeout callback. at undefined:6:28 null
+Main world window.onerror: Error: Error in user script setTimeout callback. at user-script:13:6:83 Error: Error in user script setTimeout callback.
+Main world error event listener: Error: Error in user script setTimeout callback. at user-script:13:6:83 Error: Error in user script setTimeout callback.
