+++ LayoutTests/userscripts/window-onerror-for-isolated-world-2-expected.txt
@@ -1,13 +1,13 @@
 CONSOLE MESSAGE: Error: Error in main world inline script.
 Test that window.onerror and "error" event listeners from isolated world are invoked for uncaught exceptions in user scripts running in isolate worlds as well as for exceptions in the main world.Bug 8519.
 
-user script window.onerror: Error: Error in user script inline script. at undefined:33:20 Error: Error in user script inline script.
-user script error event listener: Error: Error in user script inline script. at undefined:33:20 Error: Error in user script inline script.
+user script window.onerror: Error: Error in user script inline script. at user-script:14:33:68 Error: Error in user script inline script.
+user script error event listener: Error: Error in user script inline script. at user-script:14:33:68 Error: Error in user script inline script.
 user script window.onerror: Error: Error in main world load handler. at window-onerror-for-isolated-world-2.html:27:72 null
 user script error event listener: Error: Error in main world load handler. at window-onerror-for-isolated-world-2.html:27:72 null
-user script window.onerror: Error: Error in user script load handler. at undefined:30:24 Error: Error in user script load handler.
-user script error event listener: Error: Error in user script load handler. at undefined:30:24 Error: Error in user script load handler.
+user script window.onerror: Error: Error in user script load handler. at user-script:14:30:72 Error: Error in user script load handler.
+user script error event listener: Error: Error in user script load handler. at user-script:14:30:72 Error: Error in user script load handler.
 user script window.onerror: Error: Error in main world setTimeout callback. at window-onerror-for-isolated-world-2.html:25:83 null
 user script error event listener: Error: Error in main world setTimeout callback. at window-onerror-for-isolated-world-2.html:25:83 null
-user script window.onerror: Error: Error in user script setTimeout callback. at undefined:28:28 Error: Error in user script setTimeout callback.
-user script error event listener: Error: Error in user script setTimeout callback. at undefined:28:28 Error: Error in user script setTimeout callback.
+user script window.onerror: Error: Error in user script setTimeout callback. at user-script:14:28:83 Error: Error in user script setTimeout callback.
+user script error event listener: Error: Error in user script setTimeout callback. at user-script:14:28:83 Error: Error in user script setTimeout callback.
