+++ Source/WebKit/UIProcess/API/APIUserScript.h
@@ -34,8 +34,6 @@ namespace API {
 
 class UserScript final : public ObjectImpl<Object::Type::UserScript>, public Identified<UserScript> {
 public:
-    static WTF::URL generateUniqueURL();
-
     static Ref<UserScript> create(WebCore::UserScript&& userScript, API::ContentWorld& world)
     {
         return adoptRef(*new UserScript(WTFMove(userScript), world));
